backend/requirements.txt
fastapi
uvicorn
google-generativeai
langchain
langchain-community
pypdf
chromadb
python-multipart
python-dotenv